# openapi.json
{"openapi":"3.1.0","info":{"title":"Movie Recommendation ML Service","description":"Content-Based Filtering menggunakan TF-IDF + Cosine Similarity - Pertemuan 11","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/predict":{"post":{"summary":"Predict","operationId":"predict_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/batch-predict":{"post":{"summary":"Batch Predict","operationId":"batch_predict_batch_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchPredictRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchPredictResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BatchPredictRequest":{"properties":{"titles":{"items":{"type":"string"},"type":"array","title":"Titles"},"top_n":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Top N","default":5}},"type":"object","required":["titles"],"title":"BatchPredictRequest"},"BatchPredictResponse":{"properties":{"results":{"items":{"$ref":"#/components/schemas/PredictResponse"},"type":"array","title":"Results"},"total_processed":{"type":"integer","title":"Total Processed"},"service":{"type":"string","title":"Service","default":"python-ml-fastapi"}},"type":"object","required":["results","total_processed"],"title":"BatchPredictResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MovieResult":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"overview":{"type":"string","title":"Overview"},"genres":{"type":"string","title":"Genres"}},"type":"object","required":["id","title","overview","genres"],"title":"MovieResult"},"PredictRequest":{"properties":{"title":{"type":"string","title":"Title"},"top_n":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Top N","default":5}},"type":"object","required":["title"],"title":"PredictRequest"},"PredictResponse":{"properties":{"input_title":{"type":"string","title":"Input Title"},"recommendations":{"items":{"$ref":"#/components/schemas/MovieResult"},"type":"array","title":"Recommendations"},"total":{"type":"integer","title":"Total"},"service":{"type":"string","title":"Service","default":"python-ml-fastapi"}},"type":"object","required":["input_title","recommendations","total"],"title":"PredictResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}